move graphql data types
